pyproject.toml
[project]
name = "attendee"
version = "0.0.0"
requires-python = ">=3.11"
dependencies = [
    "amqp==5.2.0",
    "asgiref==3.8.1",
    "attrs==24.2.0",
    "azure-identity==1.25.1",
    "azure-storage-blob==12.26.0",
    "billiard==4.2.1",
    "boto3==1.35.64",
    "botocore==1.35.64",
    "cachetools==5.5.1",
    "celery==5.4.0",
    "certifi==2024.8.30",
    "cffi==1.17.1",
    "charset-normalizer==3.4.0",
    "click==8.1.7",
    "click-didyoumean==0.3.1",
    "click-plugins==1.1.1",
    "click-repl==0.3.0",
    "cryptography==43.0.3",
    "cython",
    "deepgram-sdk==4.3.0",
    "dj-database-url==2.3.0",
    "django==5.1.14",
    "django-allauth==65.1.0",
    "django-concurrency==2.6.0",
    "django-extensions==3.2.3",
    "django-storages==1.14.4",
    "djangorestframework==3.15.2",
    "docker==7.1.0",
    "drf-spectacular==0.27.2",
    "durationpy==0.9",
    "gdown",
    "google-api-core==2.24.1",
    "google-auth==2.38.0",
    "google-cloud-compute==1.32.0",
    "google-cloud-texttospeech==2.24.0",
    "googleapis-common-protos==1.66.0",
    "grpcio==1.70.0",
    "grpcio-status==1.70.0",
    "gunicorn==23.0.0",
    "h11==0.16.0",
    "idna==3.10",
    "inflection==0.5.1",
    "jmespath==1.0.1",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "jsonschema==4.23.0",
    "jsonschema-specifications==2024.10.1",
    "kombu==5.4.2",
    "kubernetes==32.0.0",
    "modal==1.4.1.dev1",
    "msgpack==1.1.0",
    "numpy==2.1.3",
    "oauthlib==3.2.2",
    "opencv-python==4.10.0.84",
    "outcome==1.3.0.post0",
    "packaging==24.2",
    "prompt-toolkit==3.0.48",
    "proto-plus==1.26.0",
    "protobuf==5.29.6",
    "psycopg2==2.9.10; sys_platform == 'linux' and platform_machine == 'x86_64'",
    "pyasn1==0.6.2",
    "pyasn1-modules==0.4.1",
    "pycparser==2.22",
    "pydub==0.25.1",
    "pygobject==3.50.0; sys_platform == 'linux' and platform_machine == 'x86_64'",
    "pyjwt==2.9.0",
    "pysaml2==7.5.4",
    "pysocks==1.7.1",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.0.1",
    "python-json-logger==4.0.0",
    "pyvirtualdisplay==3.0",
    "pyyaml==6.0.2",
    "redis==5.2.0",
    "referencing==0.35.1",
    "requests==2.32.3",
    "requests-file==2.1.0",
    "requests-oauthlib==2.0.0",
    "rpds-py==0.21.0",
    "rsa==4.9",
    "ruff==0.9.6",
    "s3transfer==0.10.3",
    "selenium==4.27.1",
    "sentry-sdk[celery,django]>=2.0.0",
    "setuptools==78.1.1",
    "six==1.16.0",
    "sniffio==1.3.1",
    "sortedcontainers==2.4.0",
    "sqlparse==0.5.1",
    "stripe==11.6.0",
    "tldextract==5.3.0",
    "trio==0.28.0",
    "trio-websocket==0.11.1",
    "typing-extensions==4.13.0",
    "tzdata==2024.2",
    "uritemplate==4.1.1",
    "urllib3==2.6.3",
    "vine==5.1.0",
    "watchdog==6.0.0",
    "wcwidth==0.2.13",
    "webrtcvad==2.0.10",
    "websocket-client==1.8.0",
    "websockets==14.1",
    "whitenoise==6.8.2",
    "wsproto==1.2.0",
    "zoom-meeting-sdk==0.0.27; sys_platform == 'linux' and platform_machine == 'x86_64'",
]

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
    "migrations",
]

# Same as Black.
line-length = 999  # Setting a very large line length to prevent wrapping

# Use Python 3.11+
target-version = "py311"

[tool.ruff.lint]
# Enable pycodestyle (`E`), Pyflakes (`F`), and import sorting (`I`)
select = ["E", "F", "I"]
ignore = ["E501", "E722", "E402", "F403"]

# Allow autofix for all enabled rules (when `--fix`) is provided.
fixable = ["A", "B", "C", "D", "E", "F", "I"]
unfixable = []

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.lint.mccabe]
# Unlike Flake8, default to a complexity level of 10.
max-complexity = 10

[tool.ruff.lint.isort]
known-first-party = ["attendee"]

[tool.ruff.format]
line-ending = "auto"
quote-style = "double"
indent-style = "space"
